ViewModels now call methods in WalletController instead of sending messages.
Deleted: SignTransactionAndShowInformationMessage
